Move and deduplicate file and checksum download methods to general blueprint
