+++ research/static/research/js/dlgFileBrowseOperations.js
@@ -125,7 +125,7 @@ $(document).ready(function () {
       } else {
         // Single file
         const filepath = `${folderPath}/${item.name}`
-        const url = '/research/browse/download?filepath=' + encodeURIComponent(filepath)
+        const url = '/browse/download?filepath=' + encodeURIComponent(filepath)
         downloadEntries.push({ url, name: folderName + '/' + item.name })
       }
     }
@@ -146,7 +146,7 @@ $(document).ready(function () {
       if (type === 'coll') {
         downloadEntries.push(...(await collectFolderEntries(path))) // reuse folder download func
       } else {
-        const url = '/research/browse/download?filepath=' + encodeURIComponent(path)
+        const url = '/browse/download?filepath=' + encodeURIComponent(path)
         downloadEntries.push({ url, name })
       }
     }
@@ -745,7 +745,7 @@ const tableRenderer2 = {
     const ext = row.name.replace(/.*\./, '').toLowerCase()
 
     const actions = $('<ul class="dropdown-menu">')
-    actions.append(`<li><a href="browse/download?filepath=${encodeURIComponent(dlgCurrentFolder + '/' + row.name)}">Download</a>`)
+    actions.append(`<li><a href="/browse/download?filepath=${encodeURIComponent(dlgCurrentFolder + '/' + row.name)}">Download</a>`)
 
     // Generate dropdown "view" actions for different media types.
     for (const type of Object.keys(viewExts).filter(type => (viewExts[type].includes(ext)))) { actions.append(`<li><a class="view-${type}" data-path="${Yoda.htmlEncode(dlgCurrentFolder + '/' + row.name)}">View</a>`) }
